refactor: use std mutex to share Settings
Previous Tokio sync usage was useful for async functions, but I did not
discover a means to access those locks in a synchronous context. With
ui view functions being synchronous functions, that proved problematic.

Switching to a std mutex allows us to access the data in a synchronous
context. The caveat being that we cannot hold a lock across an `await`,
which at the moment seems easy enough to avoid.